django_model/requirements.txt
absl-py==0.8.1
aiomysql==0.0.20
alembic==1.3.1
alipay==0.7.4
amqp==2.5.2
anaconda==0.0.1.1
aniso8601==8.0.0
anyjson==0.3.3
api==0.0.7
Appium-Python-Client==0.48
APScheduler==3.6.3
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
autobahn==20.4.3
Automat==0.8.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
baidu-aip==2.2.15.0
beautifulsoup4==4.8.1
billiard==3.6.3.0
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cachetools==3.1.1
celery==4.4.2
celery-with-redis==3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==6.7
clickclick==1.2.2
colorama==0.4.1
colorlog==4.0.2
CondConfigParser==1.0.3
configparser==4.0.2
connexion==1.1.15
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.14
daphne==2.5.0
DBUtils==1.3
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.5.0
Django==2.2.1
django-celery==3.3.1
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-mdeditor==0.1.18
django-redis==4.10.0
django-reversion==3.0.4
django-uuslug==1.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docopt==0.6.2
dpath==1.4.2
dwebsocket==0.5.10
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.25.1
example==0.1.0
fake-useragent==0.1.11
fakeredis==1.2.0
ffmpeg==1.4
filetype==1.0.5
Flask==1.1.1
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
flask-core==2.9.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
flower==0.9.4
fonttools==4.0.2
future==0.15.2
gast==0.2.2
get==2019.4.13
gevent==1.4.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.25.0
h11==0.7.0
h2==3.1.1
h5py==2.10.0
happybase==1.1.0
har2case==0.3.1
hpack==3.0.0
html2text==2020.1.16
httplib2==0.9.2
httprunner==2.2.4
humanize==0.5.1
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.7
imageio==2.5.0
imageio-ffmpeg==0.3.0
imbalanced-learn==0.5.0
imblearn==0.0
importlib-metadata==0.23
imutils==0.5.3
incremental==17.5.0
inflection==0.3.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.39
Jinja2==2.10.3
joblib==0.14.0
json-spec==0.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kaitaistruct==0.8
kazoo==2.7.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.8
ldap3==2.5.2
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.3.0
matplotlib==3.1.1
mistune==0.8.4
mitmproxy==4.0.4
monotonic==1.5
more-itertools==7.2.0
MouseInfo==0.0.4
moviepy==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==4.4.0
nose==1.3.7
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.0
opencv-python==4.1.0.25
openpyxl==3.0.0
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.0
parso==0.5.1
passlib==1.7.1
pdfminer==20191125
pickleshare==0.7.5
pika==0.13.0
Pillow==6.1.0
ply==3.11
post==2019.4.13
pqi==2.0.6
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.15
protobuf==3.11.1
psutil==5.6.7
psycopg2==2.8.3
psycopg2-binary==2.8.3
public==2019.4.13
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAutoGUI==0.9.47
pycparser==2.19
pycryptodome==3.9.7
pycryptodomex==3.7.2
PyDispatcher==2.0.5
pydivert==2.1.0
pyecharts==1.0.0
pygame==1.9.6
PyGetWindow==0.0.7
Pygments==2.4.2
PyHamcrest==1.9.0
pyipip==0.1.1
PyJWT==1.7.1
pymongo==3.8.0
PyMsgBox==1.0.7
pymssql==2.1.4
PyMySQL==0.9.2
pyodbc==4.0.30
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyperclip==1.6.5
pypinyin==0.37.0
pyquery==1.4.1
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.22
pytesseract==0.3.0
python-alipay-sdk==1.10.1
python-dateutil==2.8.1
python-docx==0.8.7
python-editor==1.0.4
python-nmap==0.6.1
python-slugify==4.0.0
PyTweening==1.0.3
pytz==2019.3
pywin32==225
pywinpty==0.5.5
PyYAML==4.2b1
pyzmq==18.1.1
qrcode==5.1
qtconsole==4.5.5
query-string==2019.4.13
queuelib==1.5.0
redis==3.3.8
request==1.0.2
requests==2.19.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
retry==0.9.2
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.15.100
schedule==0.5.0
scikit-learn==0.21.3
scipy==1.3.2
Scrapy==1.8.0
scrapy-redis==0.6.8
selenium==3.14.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.13.0
sklearn==0.0
snapshot-selenium==0.0.2
sortedcontainers==2.0.5
soupsieve==1.9.5
speaklater==1.3
SQLAlchemy==1.3.11
sqlparse==0.3.0
stomp.py==4.1.22
swagger-py-codegen==0.4.0
swagger-spec-validator==2.4.3
tablib==0.13.0
tabula-py==2.1.0
tensorboard==1.11.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
tesseract==0.1.3
testpath==0.4.4
text-unidecode==1.3
threadpool==1.3.2
thriftpy==0.3.9
tornado==5.1.1
tqdm==4.35.0
traitlets==4.3.3
Twisted==19.7.0
txaio==20.4.1
tzlocal==2.0.0
urllib3==1.23
urwid==2.0.1
video2chars==0.6.5
vine==1.3.0
virtualenv==16.7.5
w3lib==1.21.0
wcwidth==0.1.7
webargs==5.5.2
webencodings==0.5.1
websocket==0.2.1
Werkzeug==0.16.0
Whoosh==2.7.4
WhooshAlchemy==0.3.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wsproto==0.11.0
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
xpinyin==0.5.6
you-get==0.4.1355
zipp==0.6.0
zope.interface==4.6.0
